requirements.txt
alembic==1.13.1
async-timeout==4.0.3
aiohttp==3.10.3
altgraph==0.17.4
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.8
dnspython==2.6.1
fastapi==0.111.0
fastapi-cli==0.0.4
flower==2.0.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanize==4.9.0
idna==3.7
Jinja2==3.1.4
kombu==5.3.7
ldap3==2.9.1
loguru==0.7.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.0
orjson==3.10.5
packaging==24.1
pefile==2023.2.7
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rich==13.7.1
setuptools==71.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
tornado==6.4.1
transliterate==1.10.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uvicorn==0.30.1
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
win32-setctime==1.1.0